+++ packages/drawer/src/Drawer/index.ts
@@ -1,4 +1,3 @@
 export { Drawer } from './Drawer';
-export { MOBILE_BREAKPOINT } from './Drawer.constants';
 export { drawerClassName } from './Drawer.styles';
-export { DisplayMode, type DrawerProps } from './Drawer.types';
+export { DisplayMode, type DrawerProps, Size } from './Drawer.types';