+++ packages/drawer/src/LayoutComponent/EmbeddedDrawerLayout/EmbeddedDrawerLayout.types.ts
@@ -1,0 +1,3 @@
+import { LayoutComponentProps } from '../../LayoutComponent/LayoutComponent.types';
+
+export type EmbeddedDrawerLayoutProps = Omit<LayoutComponentProps, 'drawer'>;