+++ chat/chat-window/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lg-chat/chat-window
 
+## 4.1.0
+
+### Minor Changes
+
+- bb848de: [LG-5391](https://jira.mongodb.org/browse/LG-5391): add variant support for when the parent `LeafyGreenChatProvider` has `variant="compact"`
+
 ## 4.0.4
 
 ### Patch Changes